test: Migrated root smoke tests to `node:test`
