+++ test/smoke/proxy-api-connection-noproxy.test.js
@@ -4,15 +4,15 @@
  */
 
 'use strict'
-
-const tap = require('tap')
+const test = require('node:test')
+const assert = require('node:assert')
 const configurator = require('../../lib/config')
 const Agent = require('../../lib/agent')
 const CollectorAPI = require('../../lib/collector/api')
 const { getTestSecret } = require('../helpers/secrets')
 
 const license = getTestSecret('TEST_LICENSE')
-tap.test('no proxy set should not use proxy agent', (t) => {
+test('no proxy set should not use proxy agent', (t, end) => {
   const config = configurator.initialize({
     app_name: 'node.js Tests',
     license_key: license,
@@ -34,18 +34,17 @@ tap.test('no proxy set should not use proxy agent', (t) => {
   const api = new CollectorAPI(agent)
 
   api.connect((error, response) => {
-    t.notOk(error, 'connected without error')
+    assert.ok(!error, 'connected without error')
 
     const returned = response && response.payload
-    t.ok(returned, 'got boot configuration')
-    t.ok(returned.agent_run_id, 'got run ID')
-    t.ok(agent.config.run_id, 'run ID set in configuration')
+    assert.ok(returned, 'got boot configuration')
+    assert.ok(returned.agent_run_id, 'got run ID')
+    assert.ok(agent.config.run_id, 'run ID set in configuration')
 
     api.shutdown((error) => {
-      t.notOk(error, 'should have shut down without issue')
-      t.notOk(agent.config.run_id, 'run ID should have been cleared by shutdown')
-
-      t.end()
+      assert.ok(!error, 'should have shut down without issue')
+      assert.ok(!agent.config.run_id, 'run ID should have been cleared by shutdown')
+      end()
     })
   })
 })
